% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{fgcz_use_template}
\alias{fgcz_use_template}
\title{Copy the starter template into a directory}
\usage{
fgcz_use_template(dir, to = "template.qmd", overwrite = FALSE)
}
\arguments{
\item{dir}{Destination directory. Created if it does not exist.}

\item{to}{Filename for the copied template within \code{dir}.}

\item{overwrite}{Overwrite an existing file of the same name.}
}
\value{
Path to the copied template, invisibly.
}
\description{
Copies \code{template.qmd} (a generic FGCZ report skeleton) and its
\code{fgcz-report-overview.svg} visual abstract into \code{dir}, together with the
styling assets they need. Use this to bootstrap a new report.
}
\examples{
\dontrun{
fgcz_use_template("my_report", to = "my_report.qmd")
}
}
